test(input): migrate card
