+++ core/src/components/input/test/card/input.e2e.ts-snapshots/input-card-md-ltr-Mobile-Firefox-linux.png
